cleaned up python code
